Cleanup UIKitExtensions structure
