+++ Cartfile.resolved
@@ -1,9 +1,9 @@
-github "ReactiveCocoa/ReactiveCocoa" "7.0.1"
-github "ReactiveCocoa/ReactiveSwift" "3.0.0"
+github "ReactiveCocoa/ReactiveCocoa" "7.1.0"
+github "ReactiveCocoa/ReactiveSwift" "3.1.0"
 github "antitypical/Result" "3.2.4"
 github "ashleymills/Reachability.swift" "v4.1.0"
-github "realm/realm-cocoa" "v3.0.2"
-github "rs/SDWebImage" "4.2.2"
+github "realm/realm-cocoa" "v3.1.1"
+github "rs/SDWebImage" "4.3.0"
 github "thoughtbot/Argo" "v4.1.2"
 github "thoughtbot/Runes" "v4.1.0"
 github "utahiosmac/Marshal" "1.2.4"
